requirements.txt
# Use Python 3.10, 3.11, or 3.12. CatBoost has no wheel for 3.14 and will fail to build.
fastapi
uvicorn
pandas
numpy
joblib
shap
catboost
lightgbm
openai
google-generativeai
